+++ consensus/polybft/mocks_test.go
@@ -108,8 +108,8 @@ func (m *blockchainMock) GetHeaderByHash(hash types.Hash) (*types.Header, bool) 
 	panic("Unsupported mock for GetHeaderByHash") //nolint:gocritic
 }
 
-func (m *blockchainMock) GetSystemState(config *PolyBFTConfig, provider contract.Provider) SystemState {
-	args := m.Called(config, provider)
+func (m *blockchainMock) GetSystemState(provider contract.Provider) SystemState {
+	args := m.Called(provider)
 
 	return args.Get(0).(SystemState) //nolint:forcetypeassert
 }
