+++ e2e-polybft/e2e/bridge_test.go
@@ -1,13 +1,8 @@
 package e2e
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"math/big"
-	"net/http"
 	"path"
 	"strconv"
 	"strings"
@@ -33,28 +28,6 @@ import (
 const (
 	manifestFileName = "manifest.json"
 )
-
-// checkStateSyncResultLogs is helper function which parses given StateSyncResultEvent event's logs,
-// extracts status topic value and makes assertions against it.
-func checkStateSyncResultLogs(
-	t *testing.T,
-	logs []*ethgo.Log,
-	expectedCount int,
-) {
-	t.Helper()
-	require.Len(t, logs, expectedCount)
-
-	var stateSyncResultEvent contractsapi.StateSyncResultEvent
-	for _, log := range logs {
-		doesMatch, err := stateSyncResultEvent.ParseLog(log)
-		require.True(t, doesMatch)
-		require.NoError(t, err)
-
-		t.Logf("Block Number=%d, Decoded Log=%+v\n", log.BlockNumber, stateSyncResultEvent)
-
-		require.True(t, stateSyncResultEvent.Status)
-	}
-}
 
 func TestE2E_Bridge_DepositAndWithdrawERC20(t *testing.T) {
 	const (
@@ -582,92 +555,6 @@ func TestE2E_Bridge_L2toL1ExitMultiple(t *testing.T) {
 			require.True(t, isProcessed)
 		}
 	}
-}
-
-func sendExitTransaction(
-	sidechainKey *ethgow.Key,
-	rootchainKey ethgo.Key,
-	proof types.Proof,
-	checkpointBlock uint64,
-	stateSenderData []byte,
-	l1ExitTestAddr,
-	exitHelperAddr ethgo.Address,
-	l1TxRelayer txrelayer.TxRelayer,
-	exitEventID uint64) (bool, error) {
-	var exitEventAPI contractsapi.L2StateSyncedEvent
-
-	proofExitEventEncoded, err := exitEventAPI.Encode(&polybft.ExitEvent{
-		ID:       exitEventID,
-		Sender:   sidechainKey.Address(),
-		Receiver: l1ExitTestAddr,
-		Data:     stateSenderData,
-	})
-	if err != nil {
-		return false, err
-	}
-
-	leafIndex, ok := proof.Metadata["LeafIndex"].(float64)
-	if !ok {
-		return false, fmt.Errorf("could not get leaf index from exit event proof. Leaf from proof: %v", proof.Metadata["LeafIndex"])
-	}
-
-	receipt, err := ABITransaction(l1TxRelayer, rootchainKey, contractsapi.ExitHelper, exitHelperAddr,
-		"exit",
-		big.NewInt(int64(checkpointBlock)),
-		uint64(leafIndex),
-		proofExitEventEncoded,
-		proof.Data,
-	)
-
-	if err != nil {
-		return false, err
-	}
-
-	if receipt.Status != uint64(types.ReceiptSuccess) {
-		return false, errors.New("transaction execution failed")
-	}
-
-	return isExitEventProcessed(exitEventID, exitHelperAddr, l1TxRelayer)
-}
-
-func getExitProof(rpcAddress string, exitID uint64) (types.Proof, error) {
-	query := struct {
-		Jsonrpc string   `json:"jsonrpc"`
-		Method  string   `json:"method"`
-		Params  []string `json:"params"`
-		ID      int      `json:"id"`
-	}{
-		"2.0",
-		"bridge_generateExitProof",
-		[]string{fmt.Sprintf("0x%x", exitID)},
-		1,
-	}
-
-	d, err := json.Marshal(query)
-	if err != nil {
-		return types.Proof{}, err
-	}
-
-	resp, err := http.Post(rpcAddress, "application/json", bytes.NewReader(d))
-	if err != nil {
-		return types.Proof{}, err
-	}
-
-	s, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return types.Proof{}, err
-	}
-
-	rspProof := struct {
-		Result types.Proof `json:"result"`
-	}{}
-
-	err = json.Unmarshal(s, &rspProof)
-	if err != nil {
-		return types.Proof{}, err
-	}
-
-	return rspProof.Result, nil
 }
 
 func TestE2E_Bridge_ChangeVotingPower(t *testing.T) {
